requirement.txt
asgiref==3.5.2
Django==4.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
mysql==0.0.3
mysqlclient==2.1.1
PyJWT==2.4.0
python-dotenv==0.21.0
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
typing==3.7.4.3
tzdata==2022.2
